+++ Revo.Infrastructure/Security/Commands/IEntityQueryAuthorizer.cs
@@ -9,6 +9,9 @@ namespace Revo.Infrastructure.Security.Commands
     public interface IEntityQueryAuthorizer
     {
         Task<IQueryable<T>> AuthorizeQueryAsync<T, TResult>(IQueryable<T> results,
+            Expression<Func<T, TResult>> authorizedEntity);
+
+        Task<IQueryable<T>> AuthorizeQueryAsync<T, TResult>(IQueryable<T> results,
             Expression<Func<T, TResult>> authorizedEntity, ICommandBase command);
     }
 }