+++ Providers/EFCore/Revo.EFCore/DataAccess/Query/CustomQueryTranslationConfigurer.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Query;
+using Revo.EFCore.DataAccess.Configuration;
+
+namespace Revo.EFCore.DataAccess.Query
+{
+    public class CustomQueryTranslationConfigurer : IEFCoreConfigurer
+    {
+        public void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        {
+           // optionsBuilder
+           //     .ReplaceService<IQueryTranslationPreprocessorFactory, CustomQueryTranslationPreprocessorFactory>();
+        }
+    }
+}